fix(workspaces): better distinction between service-level guarantees and repo-level guarantees

+++ packages/server/modules/workspaces/services/workspaceRoleCreation.ts
@@ -2,9 +2,11 @@ import {
   DeleteWorkspaceRole,
   EmitWorkspaceEvent,
   GetWorkspaceRoleForUser,
+  GetWorkspaceRoles,
   UpsertWorkspaceRole
 } from '@/modules/workspaces/domain/operations'
 import { WorkspaceAcl } from '@/modules/workspaces/domain/types'
+import { isUserLastWorkspaceAdmin } from '@/modules/workspaces/utils/isUserLastWorkspaceAdmin'
 import { WorkspaceEvents } from '@/modules/workspacesCore/domain/events'
 
 type WorkspaceRoleDeleteArgs = {
@@ -14,9 +16,11 @@ type WorkspaceRoleDeleteArgs = {
 
 export const deleteWorkspaceRoleFactory =
   ({
+    getWorkspaceRoles,
     deleteWorkspaceRole,
     emitWorkspaceEvent
   }: {
+    getWorkspaceRoles: GetWorkspaceRoles
     deleteWorkspaceRole: DeleteWorkspaceRole
     emitWorkspaceEvent: EmitWorkspaceEvent
   }) =>
@@ -24,11 +28,19 @@ export const deleteWorkspaceRoleFactory =
     userId,
     workspaceId
   }: WorkspaceRoleDeleteArgs): Promise<WorkspaceAcl | null> => {
+    const workspaceRoles = await getWorkspaceRoles({ workspaceId })
+
+    if (isUserLastWorkspaceAdmin(workspaceRoles, userId)) {
+      throw new Error('Cannot remove last admin from a workspace.')
+    }
+
     const deletedRole = await deleteWorkspaceRole({ userId, workspaceId })
 
-    if (!!deletedRole) {
-      emitWorkspaceEvent({ event: WorkspaceEvents.RoleDeleted, payload: deletedRole })
+    if (!deletedRole) {
+      return null
     }
+
+    emitWorkspaceEvent({ event: WorkspaceEvents.RoleDeleted, payload: deletedRole })
 
     return deletedRole
   }
@@ -49,16 +61,26 @@ export const getWorkspaceRoleFactory =
 
 export const setWorkspaceRoleFactory =
   ({
+    getWorkspaceRoles,
     upsertWorkspaceRole,
     emitWorkspaceEvent
   }: {
+    getWorkspaceRoles: GetWorkspaceRoles
     upsertWorkspaceRole: UpsertWorkspaceRole
     emitWorkspaceEvent: EmitWorkspaceEvent
   }) =>
   async ({ userId, workspaceId, role }: WorkspaceAcl): Promise<void> => {
+    const workspaceRoles = await getWorkspaceRoles({ workspaceId })
+
+    if (
+      isUserLastWorkspaceAdmin(workspaceRoles, userId) &&
+      role !== 'workspace:admin'
+    ) {
+      throw new Error('Cannot remove last admin from a workspace.')
+    }
+
     await upsertWorkspaceRole({ userId, workspaceId, role })
 
-    // TODO: Should we return the final record from `upsert`, or `get`, instead of emitting args directly?
     await emitWorkspaceEvent({
       event: WorkspaceEvents.RoleUpdated,
       payload: { userId, workspaceId, role }
